+++ extensions/chunked/Cargo.toml
@@ -1,0 +1,18 @@
+# Copyright (C) Nitrokey GmbH
+# SPDX-License-Identifier: CC0-1.0
+
+[package]
+name = "trussed-chunked"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[dependencies]
+serde.workspace = true
+serde-byte-array.workspace = true
+trussed.workspace = true
+
+[features]
+encrypted-chunked = []
