+++ src/chunked/store.rs
