+++ src/manage.rs
@@ -1,152 +1,19 @@
 // Copyright (C) Nitrokey GmbH
 // SPDX-License-Identifier: Apache-2.0 or MIT
 
-use littlefs2::{
-    fs::DirEntry,
-    path,
-    path::{Path, PathBuf},
-};
-use serde::{Deserialize, Serialize};
+use littlefs2::{fs::DirEntry, path, path::Path};
 use trussed::{
-    serde_extensions::{Extension, ExtensionClient, ExtensionImpl, ExtensionResult},
+    serde_extensions::{Extension, ExtensionImpl},
     store::Store,
     types::Location,
     Error,
 };
+use trussed_manage::{
+    FactoryResetClientReply, FactoryResetClientRequest, FactoryResetDeviceReply,
+    FactoryResetDeviceRequest, ManageExtension, ManageReply, ManageRequest,
+};
 
 use crate::StagingBackend;
-
-pub struct ManageExtension;
-
-/// Factory reset the entire device
-///
-/// This will reset all filesystems
-#[derive(Debug, Deserialize, Serialize, Copy, Clone)]
-pub struct FactoryResetDeviceRequest;
-
-/// Factory reset a specific application
-///
-/// This will reset all data for a specific client
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub struct FactoryResetClientRequest {
-    pub client: PathBuf,
-}
-
-#[allow(clippy::large_enum_variant)]
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub enum ManageRequest {
-    FactoryResetDevice(FactoryResetDeviceRequest),
-    FactoryResetClient(FactoryResetClientRequest),
-}
-
-impl From<FactoryResetClientRequest> for ManageRequest {
-    fn from(value: FactoryResetClientRequest) -> Self {
-        Self::FactoryResetClient(value)
-    }
-}
-
-impl TryFrom<ManageRequest> for FactoryResetClientRequest {
-    type Error = Error;
-    fn try_from(value: ManageRequest) -> Result<Self, Self::Error> {
-        match value {
-            ManageRequest::FactoryResetClient(v) => Ok(v),
-            _ => Err(Error::InternalError),
-        }
-    }
-}
-
-impl From<FactoryResetDeviceRequest> for ManageRequest {
-    fn from(value: FactoryResetDeviceRequest) -> Self {
-        Self::FactoryResetDevice(value)
-    }
-}
-
-impl TryFrom<ManageRequest> for FactoryResetDeviceRequest {
-    type Error = Error;
-    fn try_from(value: ManageRequest) -> Result<Self, Self::Error> {
-        match value {
-            ManageRequest::FactoryResetDevice(v) => Ok(v),
-            _ => Err(Error::InternalError),
-        }
-    }
-}
-
-/// Factory reset the entire device
-///
-/// This will reset all filesystems
-#[derive(Debug, Deserialize, Serialize, Copy, Clone)]
-pub struct FactoryResetDeviceReply;
-
-/// Factory reset a specific application
-///
-/// This will reset all data for a specific client
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub struct FactoryResetClientReply;
-
-#[derive(Debug, Deserialize, Serialize, Clone)]
-pub enum ManageReply {
-    FactoryResetDevice(FactoryResetDeviceReply),
-    FactoryResetClient(FactoryResetClientReply),
-}
-
-impl From<FactoryResetClientReply> for ManageReply {
-    fn from(value: FactoryResetClientReply) -> Self {
-        Self::FactoryResetClient(value)
-    }
-}
-
-impl TryFrom<ManageReply> for FactoryResetClientReply {
-    type Error = Error;
-    fn try_from(value: ManageReply) -> Result<Self, Self::Error> {
-        match value {
-            ManageReply::FactoryResetClient(v) => Ok(v),
-            _ => Err(Error::InternalError),
-        }
-    }
-}
-
-impl From<FactoryResetDeviceReply> for ManageReply {
-    fn from(value: FactoryResetDeviceReply) -> Self {
-        Self::FactoryResetDevice(value)
-    }
-}
-
-impl TryFrom<ManageReply> for FactoryResetDeviceReply {
-    type Error = Error;
-    fn try_from(value: ManageReply) -> Result<Self, Self::Error> {
-        match value {
-            ManageReply::FactoryResetDevice(v) => Ok(v),
-            _ => Err(Error::InternalError),
-        }
-    }
-}
-
-impl Extension for ManageExtension {
-    type Request = ManageRequest;
-    type Reply = ManageReply;
-}
-
-type ManageResult<'a, R, C> = ExtensionResult<'a, ManageExtension, R, C>;
-
-pub trait ManageClient: ExtensionClient<ManageExtension> {
-    /// Factory reset the entire device
-    ///
-    /// This will reset all filesystems
-    fn factory_reset_device(&mut self) -> ManageResult<'_, FactoryResetDeviceReply, Self> {
-        self.extension(FactoryResetDeviceRequest)
-    }
-
-    /// Factory reset the entire client
-    ///
-    fn factory_reset_client(
-        &mut self,
-        client: &Path,
-    ) -> ManageResult<'_, FactoryResetClientReply, Self> {
-        self.extension(FactoryResetClientRequest {
-            client: client.into(),
-        })
-    }
-}
 
 #[derive(Debug, Clone)]
 pub struct State {
@@ -179,8 +46,6 @@ fn callback(
 ) -> impl Fn(&DirEntry) -> bool {
     move |f| !should_preserve_file(f.path(), location)
 }
-
-impl<C: ExtensionClient<ManageExtension>> ManageClient for C {}
 
 impl ExtensionImpl<ManageExtension> for StagingBackend {
     fn extension_request<P: trussed::Platform>(
